moved even more code pieces into their own module
